refactor: modularize RAG components
