+++ vgj_chat/data/__init__.py
@@ -1,0 +1,3 @@
+"""Data utilities for VGJ Chat."""
+
+__all__ = []
